pyproject.toml
[project]
name = "py-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.13"
dependencies = [
    "fastapi>=0.135.1",
    "langchain>=1.2.12",
    "langchain-openai>=0.2.0",
    "uvicorn[standard]>=0.32.0",
    "pyyaml>=6.0",
    "httpx>=0.28.0",
    "pillow>=10.0.0",
]
